Add Provisioning DXE driver for manufacturing (#18)
This patch adds a Provisioning driver that's used to prepare a device
during manufacturing. It is responsible for the following operations:

1) Create an fTPM Endorsement Key Certificate with default RSA template.
2) Load the fTPM EK Cert and send it to a provisioning host.
3) Receive a cross-signed EK Cert from the provisioning host and save it
into UEFI variables.
4) Receive per-device SMBIOS customizations from the provisioning host
and save them into UEFI variables.
5) Store a "DeviceProvisioned" key in UEFI variables so the driver
returns immediately on subsequent boots.

The driver will attempt to contact the provisioning host and if it does
not receive a response within 5 seconds it will continue boot without
provisioning the device.

Variables are saved in namespace {72096f5b-2ac7-4e6d-a7bb-bf947d673415}
This is so that the SMBIOS driver can retrieve these values and the
Operating System can query and populate these values as necessary.
Note that FakeSMBIOSDataInVolatileStorage in SMBIOS is called to aid in testing
until non-volatile UEFI variables are enabled.

Previously EfiTimeToEpoch was failing due to the Virtual RTC starting at
year 0 month 0 day 0. Update the starting point to be Jan 1 1970.

Signed-off-by: Henry Beberman <hebeberm@microsoft.com>